requirements.txt
python-telegram-bot==22.5
openai==1.51.0
python-dotenv==1.0.1